+++ include/psen_scan_v2/laserscan_conversions.h
@@ -22,7 +22,7 @@
 
 namespace psen_scan_v2
 {
-static LaserScan toLaserScan(const MonitoringFrameMsg& frame)
+static LaserScan toLaserScan(const monitoring_frame::Message& frame)
 {
   const auto resolution = frame.resolution();
   const auto min_angle = frame.fromTheta();
@@ -31,6 +31,7 @@ static LaserScan toLaserScan(const MonitoringFrameMsg& frame)
 
   LaserScan scan(resolution, min_angle, max_angle);
   scan.setMeasurements(frame.measures());
+  scan.setIntensities(frame.intensities());
 
   return scan;
 }