+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package psen_scan_v2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Forthcoming
+-----------
+* Add intensity data to LaserScan
+* Contributors: Pilz GmbH and Co. KG
 
 0.1.0 (2020-10-19)
 ------------------